Added quite a few more options to the options panel, including:  frame height and width  frame background opacity  damage type bar height
Added a "health bar" background, if you'd like to visually see how much health it will heal
Added the ability to display the number as a percent of your max health
Added the ability to hide the number value entirely, if you'd rather just use the health bar style

I would like to smooth out some minor UI issues with the sliders before this becomes a "release", as they can behave a bit oddly. Perhaps I'll need to add text boxes that allow you to fill in exact values, for when they misbehave.