Added a text indication of the known wildflesh/claws of ursoc bonus to the options menu.
Fixed a bug where claws of ursoc bonus wasn't being applied.

Attempted to make the sliders only step to integer values, but it seems theres a bug that makes this suprisingly difficult. The accepted workaround is to use a callback to round the number and feed it back to the slider, but it seems LUA is extremely terrible at rounding numbers. If I use math.ceil, numbers jump around - 140 jumps straight to 190, skipping everything in between. math.floor doesn't seem to do anything, as decimals still remain. Formatting it as a string and then back as a number seems to work most of the time, except some numbers still have minute decimals on them for some reason. I've wasted 2 hours trying to get numbers displayed on a screen properly, and this is where I give up. In the future I will replace the sliders with text boxes and buttons, but I'd like to push this release out sooner than that.